refactor(tsr-rsc-spa): organize code
