+++ tsr-rsc-spa/src/framework/entry.rsc.tsx
@@ -1,12 +1,8 @@
 import * as ReactServer from "@vitejs/plugin-rsc/rsc";
 import { ServerError } from "./error/server";
-// import type { ReactFormState } from 'react-dom/client'
-// import { Root } from '../root.tsx'
 
 export type RscPayload = {
   root: React.ReactNode;
-  // returnValue?: unknown
-  // formState?: ReactFormState
 };
 
 export type RscRequestMeta = {
@@ -15,70 +11,51 @@ export type RscRequestMeta = {
 };
 
 export default async function handler(request: Request): Promise<Response> {
-  // TODO: server action?
-  // const isAction = request.method === 'POST'
-  // let returnValue: unknown | undefined
-  // let formState: ReactFormState | undefined
-  // let temporaryReferences: unknown | undefined
-  // if (isAction) {
-  //   const actionId = request.headers.get('x-rsc-action')
-  //   if (actionId) {
-  //     const contentType = request.headers.get('content-type')
-  //     const body = contentType?.startsWith('multipart/form-data')
-  //       ? await request.formData()
-  //       : await request.text()
-  //     temporaryReferences = ReactServer.createTemporaryReferenceSet()
-  //     const args = await ReactServer.decodeReply(body, { temporaryReferences })
-  //     const action = await ReactServer.loadServerAction(actionId)
-  //     returnValue = await action.apply(null, args)
-  //   } else {
-  //     const formData = await request.formData()
-  //     const decodedAction = await ReactServer.decodeAction(formData)
-  //     const result = await decodedAction()
-  //     formState = await ReactServer.decodeFormState(result, formData)
-  //   }
-  // }
+  const rscResponse = await handleRscRequest(request);
+  if (rscResponse) {
+    return rscResponse;
+  }
+  return new Response("Not Found", {
+    status: 404,
+  })
+}
 
+async function handleRscRequest(
+  request: Request,
+): Promise<Response | undefined> {
   const url = new URL(request.url);
-  if (url.pathname !== "/__rsc") {
-    return notFound();
-  }
+  if (url.pathname !== "/__rsc") return;
   const metaRaw = url.searchParams.get("meta");
-  if (!metaRaw) {
-    return notFound();
-  }
+  if (!metaRaw) return;
 
   const meta = JSON.parse(metaRaw) as RscRequestMeta;
   let root: React.ReactNode;
+  
   switch (meta.routeId) {
     case "/test": {
-      const mod = await import("../routes/-test.server");
+      const mod = await import("../routes/test.rsc");
       root = <mod.default {...(meta as any)} />;
       break;
     }
     case "/posts": {
-      const mod = await import("../routes/-posts.route.server");
+      const mod = await import("../routes/posts/route.rsc");
       root = <mod.default {...(meta as any)} />;
       break;
     }
     case "/posts/$postId": {
-      const mod = await import("../routes/-posts.$postId.server");
+      const mod = await import("../routes/posts/$postId.rsc");
       root = <mod.default {...(meta as any)} />;
       break;
     }
     default: {
-      return notFound();
+      return;
     }
   }
 
   const rscPayload: RscPayload = {
     root,
-    // root: <Root />,
-    // formState,
-    // returnValue
   };
   const rscOptions = {
-    // temporaryReferences,
     onError(e: unknown) {
       if (e instanceof ServerError) {
         return e.digest;
@@ -97,10 +74,6 @@ export default async function handler(request: Request): Promise<Response> {
       vary: "accept",
     },
   });
-}
-
-function notFound() {
-  return new Response("Not Found", { status: 404 });
 }
 
 if (import.meta.hot) {
