+++ tsr-rsc-spa/src/routes/posts/index.tsx
