+++ tsr-rsc-spa/src/routes/posts/route.tsx
@@ -1,4 +1,4 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { tsrRscRoute } from "../framework/client";
+import { tsrRscRoute } from "../../framework/client";
 
 export const Route = createFileRoute("/posts")(tsrRscRoute());
