+++ tsr-rsc-spa/src/routes/test.rsc.tsx
