+++ tsr-rsc-spa/vite.config.ts
@@ -15,7 +15,10 @@ export default defineConfig((env) => ({
         rsc: "./src/framework/entry.rsc.tsx",
       },
     }),
-    env.command === 'serve' && tanstackRouterGenerator(),
+    !env.isPreview &&
+      tanstackRouterGenerator({
+        routeFileIgnorePattern: "\\.rsc\\.",
+      }),
   ],
 }));
